Add EmailList styles
