+++ src/App.css
@@ -6,6 +6,10 @@
     box-sizing: border-box;
 }
 
+html {
+    scroll-behavior: smooth;
+}
+
 body {
     font-family: 'Varela Round', sans-serif;
     background: #f4f5f9;
@@ -20,7 +24,7 @@ button {
 }
 
 .container {
-    max-width: 1200px;
+    max-width: 1400px;
     margin: 0 auto;
 }
