+++ src/component/EmailBody/style.css
@@ -1,5 +1,5 @@
 .email-body {
-    width: 70%;
+    width: 60%;
 }
 
 .email-body .container {
